+++ chrome/js/main.js
@@ -1,6 +1,4 @@
 var hasError = false;
-var ENTER_KEY = 13;
-var cal;
 
 /**
  * Toggle the visibility of the sidebar
@@ -41,10 +39,14 @@ function generateClassOutput() {
   $(".unitOverview").html(""); // Clear what was there before
 
   // Create headings and containers for selected class data
-  $(".class_list a").each(function(index) {
-    var unitContainer = crel("div", {"id": $(this).text()}, // Unit Container
+  $(".class-list a").each(function(index) {
+    var unitContainer = crel("div", {
+        "id": $(this).text()
+      }, // Unit Container
       crel("h3", $(this).text()), // Unit Heading
-      crel("div", {"class": "selected_classes"}) // Class Container
+      crel("div", {
+        "class": "selected_classes"
+      }) // Class Container
     );
     $(".unitOverview").append(unitContainer);
   });
@@ -56,8 +58,8 @@ function generateClassOutput() {
     // Check for missing data
     if (overview.indexOf("undefined") > -1) {
       showError("Old or broken class data has been detected! " +
-      "This may be caused by an update, or by invalid HTML editing. " +
-      "Note that you should refresh the page if you fix manually.");
+        "This may be caused by an update, or by invalid HTML editing. " +
+        "Note that you should refresh the page if you fix manually.");
     }
 
     // Append selected class data to the container with the same unit ID
@@ -70,13 +72,14 @@ function generateClassOutput() {
  * Load a hint underneath the calendar
  */
 function loadHint() {
-  // TODO Load tips from an external source
+  // TODO Load hints from an external source
   var hints = [
     "Hover your mouse over a class type in the sidebar to preview all available classes of that type!",
     "Search for a unit code or description using the search bar at the top-left and pressing 'Enter'!",
     "Use the dropdown in the search bar to select your campus when searching for classes!",
     "New to QUT? We have 3 campuses: Gardens Point (GP), Kelvin Grove (KG) and Caboolture (CB)!",
-    "These tips will show every time you refresh the page. Want to see a new one? Ctrl+R!"
+    "These tips will show every time you refresh the page. Want to see a new one? Ctrl+R!",
+    "You can search for a unit code, or a unit description using the 'unit search' box in the sidebar!"
   ];
 
   // Randomly select a hint from the array
@@ -104,7 +107,6 @@ $(window).unload(function() {
 });
 
 $(document).ready(function() {
-
   jconfirm.defaults = {
     backgroundDismiss: true,
     keyboardEnabled: true,
@@ -112,45 +114,46 @@ $(document).ready(function() {
   };
 
   // Load the campus selector options for unit search
-  getAllSemesterIDs().done(function() {
+  getSemesterIDs().done(function() {
     // Load the previous campus into the dropdown
     loadCampus();
+  }).fail(function() {
+
   });
 
   // Load the helpful hints underneath the calendar
   loadHint();
 
   // Initialize the calendar
-  cal = $(".calendar").fullCalendar({
+  var cal = $(".calendar").fullCalendar({
     header: false,
     allDaySlot: false,
     allDayDefault: false,
     defaultView: "agendaWeek",
+    displayEventTime: false,
     minTime: "07:00:00",
     maxTime: "22:00:00",
     height: "auto",
-    columnFormat: { week: "ddd" },
+    columnFormat: "ddd",
     slotEventOverlap: false,
-    weekends: false,
-    editable: false,
-    droppable: false
+    weekends: false
   });
 
   // Load the class data into the sidebar and calendar
-  loadClassData();
+  loadClassData(cal);
 
   /**
    * Trigger a search when the user presses the Enter key in the search bar
    */
   $("#unit-search").keyup(function(event) {
-    if (event.keyCode == ENTER_KEY) {
+    if (event.keyCode == 13) { // ENTER_KEY
       var searchText = $(this).val().trim();
       if (searchText !== "") {
         var semesterID = $("#campus-selector").val();
 
         // Determine if search is a unit code or unit description
-        var regex = /\b[a-zA-Z]{3}\d{3}\b/; // 3 letters, 3 digits
-        if (regex.test(searchText)){
+        var unitCodeRegex = /\b[a-zA-Z]{3}\d{3}\b/; // 3 letters, 3 digits
+        if (unitCodeRegex.test(searchText)) {
           searchUnitCode(searchText, semesterID);
         } else {
           searchDescription(searchText, semesterID);
@@ -165,7 +168,7 @@ $(document).ready(function() {
   /**
    * Show or hide the classes for an imported subject
    */
-  $(document).on("click", ".class_list a", function() {
+  $(document).on("click", ".class-list a", function() {
     var currentList = $(this).next(".classes");
 
     // Determine if the current list is showing
@@ -191,7 +194,7 @@ $(document).ready(function() {
    */
   $(document).on("mouseover", ".classes b", function() {
     if (!hasError) {
-      $(this).nextAll(".class").each(function() {
+      $(this).nextAll(".class:not([selected])").each(function() {
         previewClass(cal, $(this));
       });
     }
@@ -200,10 +203,18 @@ $(document).ready(function() {
   /**
    * Preview a class on the calendar when the user hovers over it in the sidebar
    */
-  $(document).on("mouseover", ".class", function(){
+  $(document).on("mouseover", ".class", function() {
     if (!hasError) {
       previewClass(cal, $(this));
     }
+  });
+
+  /**
+   * Preview a class on the calendar when the user hovers over it in the sidebar
+   */
+  $(document).on("mouseout", ".class", function() {
+    removeClassPreview(cal, $(this));
+    cal.fullCalendar("removeEvents", "preview");
   });
 
   /**
@@ -212,6 +223,19 @@ $(document).ready(function() {
   $(document).on("mouseout", ".classes", function() {
     cal.fullCalendar("removeEvents", "preview");
   });
+
+  function getTimes(classElement) {
+    var day = classElement.attr("day");
+
+    var start = Date.parse(day + " " + classElement.attr("start"));
+    var end = Date.parse(day + " " + classElement.attr("end"));
+
+    return {
+      "day": day,
+      "start": start,
+      "end": end
+    };
+  }
 
   /**
    * Add the class to the calendar when the user clicks on it
@@ -223,56 +247,80 @@ $(document).ready(function() {
       cal.fullCalendar("removeEvents", "preview");
       addClass(cal, $(this));
 
+      // Check for overlapping classes
+      classA = getTimes($(this));
+      var title = $(this).siblings("b");
+      var selectedClasses = $(".class-container").find("div.class[selected]");
+      selectedClasses.each(function(i, c) {
+        classB = getTimes($(c));
+
+        var sameDay = (classA.day == classB.day);
+        var timeOverlap = (classA.start < classB.end &&
+          classA.end > classB.start);
+
+        if (sameDay && timeOverlap) {
+          // TODO Make sure these classes aren't meant to overlap (e.g. PR1 & PR2)
+          // TODO Overlapping class times: Show a warning to the user
+          alert('overlap');
+
+          // Break the loop so we don't keep checking classes
+          return false;
+        }
+
+      });
+
+      addDuplicateBadge($(this));
+
       // Add the class as selected
       $(this).attr("selected", "true");
-      $(this).append(crel("div", {"class": "remove_class"}, "x"));
-      generateClassOutput();
+      $(this).append(crel("div", {
+        "class": "remove-class"
+      }, "x"));
 
-      // Update error status
-      // TODO Check if class times overlap
-      var classCount = $(this).parent().children("div[selected='selected']").length;
-      if (classCount > 1) {
-        //alert("show error");
-        // TODO Add error icon
-      }
+      // Update the selected class list
+      generateClassOutput();
 
       // Track this with GA
       var GAlabel = "[" + $(this).attr("classType") + "] " +
-      $(this).attr("day") + "(" + $(this).attr("start") + "-" +
-      $(this).attr("end") + ") " + "@ " + $(this).attr("location");
+        $(this).attr("day") + "(" + $(this).attr("start") + "-" +
+        $(this).attr("end") + ") " + "@ " + $(this).attr("location");
       _gaq.push(["_trackEvent", $(this).attr("unitName"), "enrol", GAlabel]);
     }
   });
 
   /**
-   * Remove a selected class from the calendar and its 'remove_class' button
+   * Remove a selected class from the calendar and its 'remove-class' button
    */
-  $(document).on("click", ".remove_class", function() {
+  $(document).on("click", ".remove-class", function() {
     var classElement = $(this).parent();
 
-    // Only remove that class, nothing else.
-    cal.fullCalendar("removeEvents", getClassID(classElement));
+    removeClass(cal, classElement);
 
-    // Remove the class selection
-    classElement.removeAttr("selected");
+    // TODO Check if class times overlap
+
+    removeDuplicateBadge(classElement);
+
+    // Remove the 'x' label
     $(this).remove();
+
+    // Update the selected class list
     generateClassOutput();
 
     // Track this with GA
     var GAlabel = "[" + classElement.attr("classType") + "] " +
-    classElement.attr("day") + "(" + classElement.attr("start") + "-" +
-    classElement.attr("end") + ") " + "@ " + classElement.attr("location");
+      classElement.attr("day") + "(" + classElement.attr("start") + "-" +
+      classElement.attr("end") + ") " + "@ " + classElement.attr("location");
     _gaq.push(["_trackEvent", classElement.attr("unitName"), "unenrol", GAlabel]);
 
-    // Since the button is a child of the class div,
-    // return so that we don't add the class again
+    // Since the remove button is a child of the class div,
+    // return so that the click event doesn't propegate downwards
     return false;
   });
 
   /**
    * Remove a unit from the list of imported units in the sidebar
    */
-  $(document).on("click", ".remove_unit", function() {
+  $(document).on("click", ".remove-unit", function() {
     var unitHeader = $(this).parent();
 
     // Remove all classes from this unit
@@ -295,9 +343,9 @@ $(document).ready(function() {
   /**
    * Remove all children from the class data container
    */
-  $(document).on("click", ".remove_all", function() {
+  $(document).on("click", ".remove-all", function() {
     hasError = false;
-    $(".class_container").empty();
+    $(".class-container").empty();
     generateClassOutput();
   });